\name{example-new}
\alias{example-new}
\docType{data}
\title{
An example of intensity data for SNP genotyping
}
\description{
The file \code{example-new.txt} contains some signal intensity data for
testing and comparing genotype scoring algorithms
}
\format{
  This is a text file containing data on 99 SNPs for 1550 DNA
  samples. One line of data appears for each SNP, starting with the
  SNP name and followed by 1550 pairs of intensity values. There is a
  header line containing variable names, with intensities labelled
  as \code{xxxxA} and \code{xxxxB}, where \code{xxxx} is the sample
  name.
}
\details{
See the package vignette "Comparing clustering algorithms".
}
\source{These data were originally distributed with the "Illuminus"
  genotype scoring software from the Wellcome Trust
  Sanger Institute: 
  \url{http://www.sanger.ac.uk/resources/software/illuminus/}
}
\keyword{datasets}
