\name{read.plink}
\alias{read.plink}
\title{Read a PLINK binary data file as a SnpMatrix}
\description{
The package PLINK saves genome-wide association data in groups of three
files, with the extensions \code{.bed}, \code{.bim}, and
\code{.fam}. This function reads these files and creates an object of
class \code{"SnpMatrix"}
}
\usage{
read.plink(bed, bim, fam)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{bed}{The name of the
    file containing the packed binary SNP genotype data. It should have
    the extension \code{.bed}; if it doesn't, then this extension will
    be appended}
  \item{bim}{The file containing the SNP descriptions}
  \item{fam}{The file containing subject (and, possibly, family)
    identifiers. This is basically a tab-delimited "pedfile"}
}
\details{
  If the \code{bed} argument does not contain a filename with the file
  extension \code{.bed}, then this extension is appended to the
  argument. The remaining two arguments are optional; their default
  values are obtained by replacing the \code{.bed} filename extension by
  \code{.bim} and \code{.fam} respectively. See the PLINK documentation
  for the detailed specification of these files.
}
\value{
  An object of class \code{"SnpMatrix"}. The column names are the SNP
  names as listed in the \code{bim} file. If the subject identifiers in
  the \code{fam} file are not duplicated, these are used for the row
  names of the output object. But, if there are duplicated subject
  identifiers, the row names are generated by concatenating the
  family and subject identifiers (separated by the character \code{:})
}
\references{PLINK: Whole genome association analysis toolset.
  \url{http://pngu.mgh.harvard.edu/~purcell/plink/}
}
\author{David Clayton \email{david.clayton@cimr.cam.ac.uk}}
\seealso{\code{\link{write.plink}},
  \code{\link{SnpMatrix-class}}, \code{\link{XSnpMatrix-class}}}
\keyword{manip}
\keyword{IO}
\keyword{file}
\keyword{utilities}
