\title{snpAssoc-internal}
\name{snpAssoc-internal}
\alias{genotype-class}
\alias{haplotype-class}
\description{All the dirty details that don't belong elsewhere. At the
  moment just for hiding references to the \code{genotype-class} and
  \code{haplotype-class} class which are in the obsolete
  \code{genetics} package. 
}
\keyword{package}
