\name{snpStats-package}
\alias{snpStats-package}
\alias{snpStats}
\docType{package}
\title{
\packageTitle{snpStats}
}
\description{
\packageDescription{snpStats}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{snpStats}
\packageIndices{snpStats}
}
\author{
\packageAuthor{snpStats}

Maintainer: \packageMaintainer{snpStats}
}
\keyword{ package }
